+++ cmd/client/BUILD
@@ -10,14 +10,26 @@ SRCS = [
     "util.go",
 ]
 
+config_setting(
+    name = "local",
+    values = { "define": "istio=local" }
+)
+ISTIO_DEPS = [
+    "//:mixer/api/v1",
+    "//:istio/config/v1",
+    "//:istio/config/v1/aspect/listChecker"
+]
+
 DEPS = [
-    "@com_github_istio_api//:go_default_library",
     "@com_github_golang_glog//:go_default_library",
     "@com_github_golang_protobuf//ptypes:go_default_library",
     "@com_github_golang_protobuf//ptypes/timestamp:go_default_library",
     "@com_github_spf13_cobra//:go_default_library",
     "@org_golang_google_grpc//:go_default_library",
-]
+]+select ({
+":local": ["@local_istio_api" + d for d in ISTIO_DEPS],
+"//conditions:default": ["@com_github_istio_api" + d for d in ISTIO_DEPS]
+})
 
 go_binary(
     name = "mixc",
