+++ pkg/aspect/BUILD
@@ -4,7 +4,7 @@ licenses(["notice"])
 
 load("@io_bazel_rules_go//go:def.bzl", "go_prefix", "go_library")
 
-go_prefix("istio.io/mixer/aspect")
+go_prefix("istio.io/mixer/pkg/aspect")
 
 config_setting(
     name = "local",
